python/rag-embedding/pyproject.toml
[project]
name = "sqlrooms-embedding-prep"
version = "0.1.0"
description = "Prepare embeddings from markdown files and store in DuckDB"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llama-index>=0.10.0",
    "llama-index-embeddings-huggingface>=0.2.0",
    "llama-index-vector-stores-duckdb>=0.2.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]

[project.scripts]
prepare-embeddings = "prepare_embeddings:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["prepare_embeddings.py"]

[tool.uv]
dev-dependencies = []

